requirements.txt
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
graphviz==0.8.2
h5py==2.7.1
html5lib==0.9999999
Keras==2.0.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
mock==2.0.0
numpy==1.14.2
opencv-python==3.4.0.12
pbr==4.0.0
Pillow==5.1.0
protobuf==3.5.2.post1
pydot==1.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
subprocess32==3.2.7
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
Theano==1.0.1
Werkzeug==0.14.1
